requirements.txt
beautifulsoup4==4.5.3
Django==1.10.5
django-admin-views==0.8.0
django-nested-admin==3.0.12
lxml==3.7.2
python-dateutil==2.6.0
python-monkey-business==1.0.0
simplejson==3.10.0
six==1.10.0
